class moved to a separate file
